moved functions into utility files
